fix(svar2): move FORMAT resolution to un-gated record_source for query-core build
rvk.rs (un-gated) called chunk_assembler::resolve_format, but
chunk_assembler is gated behind the conversion feature, breaking
`cargo check --no-default-features` (the query-core build genvarloader
links against). Move sentinel_default, is_htslib_missing,
resolve_scalar, and resolve_format verbatim from chunk_assembler.rs to
the un-gated record_source.rs, and update all callers
(chunk_assembler.rs, vcf_list_reader.rs, rvk.rs) to the new path.
